API Deployment Docker+Poetry/pyproject.toml
[tool.poetry]
name = "app_test/"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.23.2"
langserve = {extras = ["all"], version = "^0.0.51"}
pydantic = "<2"
langchain-openai = "^0.1.1"
python-decouple = "^3.8"
faiss-cpu = "^1.8.0"
grandalf = "^0.8"

[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
